Moved method and fixed integration tests
