+++ Tests/UnitTests/Misc/XCTestCase+Extensions.swift
@@ -70,6 +70,19 @@ extension XCTestCase {
 
 }
 
+/// Similar to `XCTUnrap` but it allows an `async` closure.
+@available(iOS 13.0, macOS 10.15, tvOS 13.0, watchOS 6.2, *)
+func XCTAsyncUnwrap<T>(
+    _ expression: @autoclosure () async throws -> T?,
+    _ message: @autoclosure () -> String = "",
+    file: StaticString = #filePath,
+    line: UInt = #line
+) async throws -> T {
+    let value = try await expression()
+
+    return try XCTUnwrap(value)
+}
+
 private extension XCTestCase {
 
     func unreachable() -> Never {
